Fix deploy_artifacts.py
Currently for each plugin in the loop it appends 'gerrit.plugins' to the
group_id once for every iteration of the loop. Which means when you
upload a lot of files you end up with a group_id that looks like:
com.googlesource.gerrit.plugins.gerrit.plugins.gerrit.plugins etc.

Change-Id: I1e4384e0fc94cb0feceef832d524663f25e8afe6