its-phabricator: new build with updated its-base
Adding a star creates a ref in All-Users pointing to a git blob that
contains "star". Since this is now an update to a ref in All-Users it
triggers an its-phabricator event, and since its-base previously
expected all refs to point to commits, this caused an ERROR.

Paladox/gerrit upstream have merged a fix for its-base. I rebuilt
its-phabricator using the new its-base and can confirm that the fix
works in testing.

Change-Id: Ic7379e091061682499d639bd7d90cb519f16df3d