Fork off a custom workspace_status.py with more heuristics
In some cases, the version description of plain `git describe` is not
ideal and could be improved by implementing a few of the common
conventions of our code base. Since implementing these conventions
might slow down stamping and we do not want to slow down the
development cycle for everybody, we fork off a dedicated
workspace_status with added heuristics.

Test plan:

  1. Make sure you're in branch stable-3.1

  2. Link stable-3.1 branch of `its-base` into plugins directory.

  3. Link stable-3.1 branch of `its-phabricator` into plugins directory.

  4. Verify that

   $ python tools/workspace_status.py | grep PHAB

 looks like

   STABLE_BUILD_ITS-PHABRICATOR_LABEL 407a93f

 (The stamp is the plain, abbreviated commit hash)

  5. Verify that

   $ python tools/workspace_status_release.py | grep PHAB

 looks like

   STABLE_BUILD_ITS-PHABRICATOR_LABEL stable-3.1-0-g407a93f(its-base:stable-3.1-0-g01c400e)

 (So it contains the `its-base` version, and exposes that both are on
 `stable-3.1`)

Change-Id: Ib1681b2730cf2c443a3cb55fe6e282f6484e18de