Keep related changes list visible, if it contains visible elements
The related changes list used to hide itself, if none of the built-in
change lists contained elements. This rule ignored plugins. So if all
built-in change lists were empty, but a plugin (e.g.: `zuul` plugin)
injected content, the plugin content was invisible.

We now hide the related changes list only if we can assert that all
children are invisible. Thereby the plugin content stays visible,
even if all built-in change lists are empty.

Change-Id: Ifc941fa479ed25adb328f07f4f2a088db57d8be0