gerrit: Convert gerrit-theme to Polymer 3
Gerrit 3.4 drops html plugin support and Gerrit 3.3+ supports js modules:
https://gerrit-review.googlesource.com/c/gerrit/+/273355

This was previously done with 048d338 in April 2021 but it had to be
reverted since we ran Gerrit 3.2 at the time and it did not support
gerrit-theme.js yet.

In August 2021 we have upgraded to Gerrit 3.3 which takes the .js file
over the html file. For some reason, after our deployment got reverted,
the javascript file has been let behind and it is the one actually
being used which I have confirmed in the network view of my browser.

The files on the server:
a0b45d4f929a36b010f89b4755e95962  /var/lib/gerrit2/review_site/static/gerrit-theme.html
3f6f78618881137e26eb730d6e8fb050  /var/lib/gerrit2/review_site/static/gerrit-theme.js

Which matches:
- the html file removed by this change confirming that no change has
  been introduced meanwhile
- the javascript file from the reverted patch 048d338 and added back
  here.

Essentially, this is already deployed.

Add a bit to ensure the old useless gerrit-theme.html is removed by
Puppet.

Bug: T299877
Signed-off-by: Antoine Musso <hashar@free.fr>
Change-Id: Icd8d6748349935d8acb96b0d7a997e343ece89c6