Download upstream war with Maven
Google releases the gerrit.war to Maven central, we can thus fetch it
using Maven.  The Maven dependency-plugin lets us download the artifact
and copy it as `/gerrit.war`.  Downloading a new version of Gerrit is
thus all about:

- editing the version of com.google.gerrit:gerrit-war in pom.xml
- mvn package
- git add pom.xml gerrit.war
- git commit -m 'Gerrit vX.Y.Z'

Maven is configured to use Wikimedia Archiva, which is merely a proxy
for Maven Central:
https://wikitech.wikimedia.org/wiki/Archiva

Change-Id: If25f5f3dd733837ac1423d8ec5bbed88f470e860