pom: overwrite gerrit.war
When bumping the gerrit version and issuing `mvn package`, the
`gerrit.war` artifact might not be overwritten. The default is to
compare based on the timestamp and only overwrite if the release is
newer.  If the previous gerrit.war got pushed in the git repository
AFTER the newer gerrit.war has been released, it will not be
overwritten.

Reference:
https://maven.apache.org/plugins/maven-dependency-plugin/copy-dependencies-mojo.html#overWriteReleases

Change-Id: I2134669e6206d61fe8ef31f2cb6b5603c355d971