Document how to test a JavaScript Gerrit plugin
The Gerrit FE dev helper chrome extension lets one inject a plugin in
the browser. It has to be served from a local web server to inject cors
header, I have reused the logic we wrote for `integration/docroot`.

The README.md gives detail as to how one can setup the requirements to
test a plugin.

Bug: T214068
Change-Id: Id74400a916703406a5ed1aead8d21af1e9957703