Add unit testing with QUnit
The Gerrit plugin can be loaded in the browser with the exposed
`Gerrit` variable:

  Gerrit.install( plugin => { .. } );

For TypeScript linting I had to explicitly rely on `window.Gerrit` which
is where it is exposed by `@gerritcodereview/typescript-api`. In the
QUnit test I thus have to forge a mock `window.Gerrit` in order to be
able to require the plugin from the tests.

Split ESLint configuration since we need different rules between the
plugin that runs in the browser and the QUnit test which runs under
NodeJS:

[/.eslintrc.json]

Now only has rules matching my personal preferences such as
`comma-dangle`. It is imported in the two other following files.

[/plugins/.eslintrc.json]

Suitable for a browser and with the jsdoc plugin introduced in
I01c60f472dcc158096bcf9e38e06b0e682a042c5

[/test/.eslintrc.json]

Suitable for NodeJS and QUnit.

Change-Id: I2e80ded7ef040d0a5dbbbae128831fe73820e5a2