wm-checks-api: show processor prototype name on error
When a a BotProcessor is not yet implementing `accept` and/or `parse` an
error is thrown which is displayed in the web UI:

  Error while fetching results for wm-checks-api:
    Error: parse() not implemented in [object Object]

Converting the Processor class to a string yields `[object Object]`
instead of the class name. We instead need to retrieve the constructor
name which will give us:

  Error while fetching results for NULL:
    Error: PipelineBotProcessor.parse()

Add a test to ensure a child BotProcessor throws error containing the
class name.

Change-Id: I6f1fbe3cb6b6aac8845279e75620189c3fc976be