wm-checks-api: add support for Puppet Catalogue Compiler
`operations/puppet.git` has the script `utils/pcc.py` which lets end
user trigger a Jenkins job and would report back on the change.

It votes Verified -1/+1 and the message has the tag autogenerated:pcc-py
since 2022-12-14 I587381f3910a56441805c2544e8b09d936cdd868.

The bot three kind of messages:

1) PCC SUCCESS (.*): <url>
2) PCC FAIL (.*): <url>
3) PCC Check manually: <url>

SUCCESS are mapped assigned the INFO category. The others are ERROR.

For a SUCCESS the node status is either a NOOP or DIFF. They are not
differentiated and are both assigned INFO.

Change-Id: Iab8e84721ece6d0ef6e12eb47a5ad2c68f6e3e62