wm-zuul-status: filter out non-live item
The plugin reports:

  Error while fetching results for wm-zuul-status:
  TypeError: status.jobs[0] is undefined

The lack of jobs is due to non-live items which do not have any jobs
running, they are merely there to represent the dependency on the Zuul
status page (see T214068#8724200).

When sending a change B which depends on A, A is represented in the Zuul
status as a "non-live" item. It allows the Zuul status page to represent
the dependency:

 ▒ mediawiki/core
 │  A,1  ▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄
 │
 █ mediawiki/extensionFoo
    B,1  ████ ▒▒▒▒ ████ ████
    some-job-docker  success
    yetanother       ███▒▒▒▒
    some-other-job   failure

When browsing change A, the plugin gets a status for that change which
looks like:

  [
   {
     "active": true,
     "jobs": [],
     "live": false,
     ...
   }
  ]

The falsey `live` property indicate the change is not processed, there
is thus no point in reporting it to the Check UI: filter all items that
are non-live.

The same error shows up immediately after an item has been enqueued:
since there are no jobs running yet, the plugin is unable to determinate
the pipeline. That will be addressed in the follow up change
I402f37c3a8413206a99d979d807fac9836ac3abc.

Bug: T214068
Change-Id: Ibd42b830c236fd3bd4ecda0b9436d42a29c6ed1f