wm-zuul-status: fix items having no build
Immediately after an item has been enqueued, it has no jobs running yet
which leads to a fetch error (T214068#8724008):

 Error while fetching results for wm-zuul-status:
 TypeError: status.jobs[0] is undefined

The item has no job running yet and thus the plugin is unable to
determine the check name. That is a bug in the Zuul change status: the
pipeline should be indicated at the item level instead of a down at the
jobs level.

Until the pipeline has been determinated, the check name should be a
placeholder such as "Waiting for build".

When at least one build has started, we can then retrieve the pipeline.

Also the CheckRun name is set solely by looking at the first job which
might not have started yet and thus does not have the pipeline set:

 checkName: status.jobs[0].pipeline

To find the pipeline, crawl through all jobs to find one that has
started

Bug: T214068
Change-Id: I402f37c3a8413206a99d979d807fac9836ac3abc