wm-checks-api: parse PCC full message
On I8ed3ff5d7712569b74a23936aba43e5039b91b00 a user had a pending draft
comment then ran the pcc python script resulting in the draft comment to
be submitted. The cover message thus had an extra (1 comment), since the
bot processor only looked at the 2 line, that cause the regex to fail
and lead to a `TypeError: compiled is null`.

Change the bot processor to process the full message instead of a
hardcoded specific line. Cover that with a test.

Change-Id: I2bb2ac15affa12c5bb362bfa9bd62edec883922b