Extract and deploy upstream plugins
Since Gerrit 3.2.11 (4f29981f5) we have removed the upstream plugins
from this repository in favor of extracting them at deployment time
using:

 java -jar bin/gerrit.war init --batch --install-all-plugins

The rationale was given during code review on:
https://gerrit.wikimedia.org/r/c/operations/software/gerrit/+/699035/

When deploying a JavaScript plugin, scap replaces the plugins directory
to the deployed revision of this repository which lacks the bundled
plugins. Potential fixes would be to have scap to either:

- run the above init --install-all-plugins, but there are lock held by
  the running Gerrit which prevents init from running
- extract the zip files from the gerrit.war and checkout the one we
  might have forked (`git checkout plugins/*.jar`). And I think that
  will complicate things to have that handled via scap.

Instead adjust `mvn package` to handle the plugins extraction for us via
the maven dependency plugin `package:unpack`. It is given the gerrit war
artifact and instructed to extract `WEB-INF/plugins/*.jar` into the
`./plugins/` directory. To strip the path, this use a fileMapper which
requires the maven-dependency-plugin 3.1.2 or later.

The goal is that when one wants to deploy a JavaScript file, we can scap
deploy it without restarting Gerrit and it will reload the plugins
automatically (`plugins.checkFrequency` defaults to 1 minute.

This also mean that when we solely deploy plugins, we do not have to
restart Gerrit.

Change-Id: Ia677b06cdd7adbad88292220cc995da49dca50fc