wm-zuul-status: change pending jobs SUCCESS > INFO
A CheckResult is supposedly immutable and represent the result of a
completed build. CheckRun however can be marked as pending but I have
used it to represent a BuildSet in a Zuul pipeline while CheckResult are
abused to represent a Jenkins build.

The issue is Gerrit proposes two layers:
- CheckRun
- CheckResult

I have mapped the Zuul pipelines to CheckRun since it did not make sense
to me to show half a hundred of jobs in the UI. Ideally we would need an
extra layer in Gerrit, but for now CheckResult represents a Jenkins
build rather than a result and a build can be in progress.

This change the category of pending or building jobs from the misleading
SUCCESS to INFO.

Bug: T214068
Change-Id: Idbeece225cb957c05a002575778a8c89466bd661