wm-zuul-status: offer to reload on CI completion
Once jobs have completed, the in-progress checks will vanish and are not
displayed to the user. We need to reload the change for the
wm-checks-api plugin to process the Zuul messages and craft the check
results.

This is done by displaying an alert to the user which offers to reload
the change. The event listener is attached to the `<gr-change-view>`
element so we need to dispatch the event after some DOM crawling.

Bug: T214068
Change-Id: Ie3436d98e84845faa8e44b1ebff93e6410dc4eea