wm-checks-api: filter out Zuul start messages
`gate-and-submit` pipelines have a `start` defined to reset the Verified
label. That is also causes Zuul to add a message indicating it is
starting the jobs:

  Patch Set 1: -Verified

  Starting gate-and-submit jobs.
  https://integration.wikimedia.org/zuul/

Since the announce does not hold any job result, reject the messages to
prevent processing it. This prevents inserting an empty CheckRun with
attempt #1 for `gate-and-submit` with the actual run being added next as
attempt #2.

Bug: T348920
Change-Id: Iac4de20ce32644532841f32fdf488427d30f43da