Remap serving plugins under /r/plugins/
When injecting a plugin, the plugin URL must match the target Gerrit
expectation, namely that plugins are server under `/r/plugins` for the
production Gerrit. Else Gerrit is unable to discover the plugin name
from the URL which logs a warning in the console and cause some side
effects in the Checks API.

Also response with a 404 when a request filed does not exist, this avoid
serving the raw plugins-router.php content when a file is not existing.

In README.md, update the example URL to configure the Gerrit FE dev
helper.

Change-Id: I777284427820a9518c338e592f607eb02facff5e