Make serve:plugins emit a 404 for missing files
When a request is made to a non existing file, the PHP built-in server
sets `$_SERVER["SCRIPT_FILENAME"]` to the router handling the request
and thus set it to `plugins-router.php`. Since the file exists, we
returned its content rather than an error. If one injects a plugin in
Gerrit with a wrong URL, the browser then tries to parse the php file
as javascript.

Check whether the requested resource (SCRIPT_NAME) exists under the
document root before reading it. Fallback to emitting a 404 not found
otherwhise.

Change-Id: I9da2ec4896e1863b3c8c2dac514b8183211d9f33