Add a banner for the 2023 developer survey
The banner shows at the top of the web view and is collapsed after a few
seconds by using CSS animations.

The dismiss link allows a user to never see the banner (unless they
clear their local storage).

Disable ESLint "yoda": Expected literal to be on the right side of !==

Bug: T351109
Change-Id: Idfdf44a28c3ff1081038f54888476f6635b9086d