Let Gerrit manage light/dark theme
Gerrit 3.7 can set the theme based on the OS preference. The theme rules
are set on the html element via the ligthTheme and darkTheme classes and
Gerrit toggle between them based on user/OS preference.

Instead of reading the localstorage, stylize the lightThem and darkTheme
classes and let Gerrit handle the class being picked up.

This is a prerequisite to upgrade to Gerrit 3.8 which removes
plugin.registerStyleModule (deprecated since 3.5).

Bug: T354886
Change-Id: I9e9a63955866d215889ecd4942a9a7e7be532144