wm-pcc: add a run action
Add a single click action to comment `check experimental` and thus run
the Puppet Catalogue Compiler.

This adds a `RUN PUPPET COMPILER` action link below the commit message
next to the check chips. As it stands out, that makes it easy to
discover.  In the `Checks` tab, there is also a `Runnable` check with a
`Run` link.

If the Puppet compiler already ran from CI, the links are not shown. If
needed it can later on be added using the same action.

Bug: T363918
Change-Id: Ifc82fa7bae2f51208341af7bfa450dfa861478d6