wm-patch-demo: silently ignore errors
The patch demo service can fail and that shows a big red chipset on any
change. There are a few issues in how the errors are emitted:
* the CORS headers might not be set
* the expected json can be some PHP warning
* it still yells a HTTP 200 when an error occured

In T363355 we have considered only running the check for MediaWiki
related repositories (I3c5c0f6f1c3e5172f036e07ad8bd41a672f04ec4) but
patch demo has a larger list of repositories it can apply patches from
and we determinated we did not want to keep in sync two lists.

Instead, ignore any error that might occurs when fetching from the
service or when parsing its response. The errors are solely logged as
errors in the console but the web UI would no more show a red chip below
the commit message.

Bug: T367155
Change-Id: I9c6007845d2def571a3083fba3a4a8a952cdd28e