Handle image-diff external dependencies
The plugin has a relative dependency:

  "polymer-bridges": "file:../../polymer-bridges/"

When copied in Gerrit core tree, that fails since the relative path now
points to the parent directory.

Read the plugin `external_package.json` to retrieve the dependencies and
forge a `plugins/package.json`.  That effectively removes the relative
link since it is in devDependencies.

Restore the yarn.lock after installation. Note: the plugin lock files
are not used at all :-\

Bug: T341291
Change-Id: I0e6b27ab84dd7721cca73ec5408bf70a58e8ac67